share/mk/: install-man: Don't pipe to install(1)
install(1) doesn't work well in Darwin.  It seems to be unable to handle
</dev/stdin>.  The problem seems to be that Darwin's stdin misbehaves.

I've only workarounded the install target, and not the dist target,
since it's unlikely that one would run it in such a broken system.  But
since installing the pages is a common operation that everybody needs,
let's work around it here.

Fixes: 30c38a8bf8ae ("*.mk: Pipe to install(1)")
Link: <https://github.com/NixOS/nixpkgs/pull/300797>
Closes: <https://bugzilla.kernel.org/show_bug.cgi?id=218730>
Reported-by: Sergei Trofimovich <slyich@gmail.com>
Reported-by: Mario Rodas <https://github.com/marsam>
Cc: Eli Schwartz <eschwartz93@gmail.com>
Cc: Pádraig Brady <P@draigBrady.com>
Signed-off-by: Alejandro Colomar <alx@kernel.org>