elf.5, ld.so.8: Undeprecate DT_RPATH; explain DT_RPATH vs DT_RUNPATH
In elf.5 and ld.so.8, remove deprecation of DT_RPATH
In elf.5, amend the description of DT_RUNPATH and DT_RPATH.

Rationale:
There is no credible path towards removal of DT_RPATH.
Lots of software depend on DT_RPATH as is.
It is used e.g. for testing and in binary installers.

Cc: Joseph Myers <josmyers@redhat.com>
Cc: Mike Frysinger <vapier@gentoo.org>
Cc: <bug-binutils@gnu.org>
Signed-off-by: Joachim Wuttke <j.wuttke@fz-juelich.de>
Signed-off-by: Alejandro Colomar <alx@kernel.org>