printf.3: Rework '
As it stands, this is worse than useless
(defined as: I had to test it explicitly against strfmon() and look at
             POSIX myself to make sure the manual wasn't lying to me
 https://101010.pl/@nabijaczleweli/112694726416515899):

No-one cares if some compiler doesn't understand something from 1994.
(Careful readers will note it's been 30 years.)

"SUSv2 doesn't have this, SUSv3 does" is reefer-induced:
this figures in SUSv1 (XPG Issue 4 Version 2)
and the CHANGE HISTORY clearly notes ' was added in i4, not i4v2.

It's important that these aren't actually grouped by the thousand,
but, instead, by "what the locale says".

Each locale has two groupings: monetary and non-monetary,
and it's paramount to indicate which is which
(hi_IN  is 3,2... monetary and 3... non-monetary;
 mjw_IN is 3,2... both).

It waxes poetic for way too long about "how to set a locale".

Also replace "The Single UNIX Specification adds" with "POSIX adds".
The SUS is, on a good day, a profile of POSIX you pay 300$ for.
Maybe you could swing this back when it was an "extension"
(so, only required in the SUS profile)
but it's just shaded CX ("this is intentionally different from C")
in Issue 8.  It's really irrelevant here.

Also add this to HISTORY
(and call it something sane there;
 the Proper nomenclature is absolutely meaningless to normal people).
The most important bit for every HISTORY reader is the 1994 date.

Signed-off-by: Ahelenia Ziemiańska <nabijaczleweli@nabijaczleweli.xyz>
Message-ID: <o2vchme4dchemjo4diziac5rlmhtsaze2yi72fzo5r67umlwny@tarta.nabijaczleweli.xyz>
Signed-off-by: Alejandro Colomar <alx@kernel.org>