__riscv_flush_icache.3, riscv_flush_icache.2: Add page (and link page)
riscv_flush_icache() has been present in syscall.2, but had no page for
itself.  Add the page for this syscall.

The syscall is named riscv_flush_icache(), but the glibc function is
named __riscv_flush_icache().  To handle this, name the syscall page
riscv_flush_icache.2 and point to the libc page __riscv_flush_icache.3.

Signed-off-by: Charlie Jenkins <charlie@rivosinc.com>
Message-ID: <20240701-flush_icache-v2-1-15f7a35a0d13@rivosinc.com>
[alx: ffix]
Signed-off-by: Alejandro Colomar <alx@kernel.org>