nextup.3: wfix + tfix
The current "If x is 0" condition is a bit misleading because "is"
is not the equality test (just like when saying "x is NaN") and 0
as a FP number stands for +0, while this condition should apply to
both -0 and +0.
Replace this condition by "If x is +0 or -0".

Replace "Nan" by "NaN" (typography used everywhere else).

Cc: Dave Kemper <saint.snit@gmail.com>
Cc: "G. Branden Robinson" <branden@debian.org>
Cc: Damian McGuckin <damianm@esi.com.au>
Cc: John Gardner <gardnerjohng@gmail.com>
Signed-off-by: Vincent Lefevre <vincent@vinc17.net>
Message-ID: <20240808115610.GC2669@cventin.lip.ens-lyon.fr>
Signed-off-by: Alejandro Colomar <alx@kernel.org>