share/mk/: build-pdf-book: Rename book to "GNU/Linux Programmer's Manual"
The manual pages in this project had historically used the 5th argument
to TH, with the string "Linux Programmer's Manual".  This was showed as
the center-top title of the pages.  This was incidentally consistent
with the UNIX Programmer's Manual.

A few years ago, I removed the 5th argument to TH to use the default
string, to follow conventions.  At the same time, I put the project name
and version in the 4th argument to TH: "Linux man-pages X.Y".

When we added the scripts written by Deri to produce a PDF book, the
title of the book was "The Linux Manpage Book", and the front page said
"GNU/Linux\nTHE MAN PAGES BOOK".  For consistency with the project name,
I changed those some time ago to be title: "The Linux man-pages Book"
and front: "GNU/Linux\nTHE MAN-PAGES BOOK".

However, for consistency with the UNIX Programmer's Manual, and with the
old title used within the project, I'm now (partially) restoring the
title of the book and its front page to be both:
"GNU/Linux Programmer's Manual".
(It's not fully restorative, because it has GNU/Linux where the old
 title had just Linux, but half of our documentation is for glibc, so I
 think it's deserved.)
(I was never convinced by the old front text: why should it have the
 word "book"?)

Cc: Douglas McIlroy <douglas.mcilroy@dartmouth.edu>
Cc: "G. Branden Robinson" <branden@debian.org>
Cc: Deri James <deri@chuzzlewit.myzen.co.uk>
Cc: "Michael T. Kerrisk" <mtk.manpages@gmail.com>
Signed-off-by: Alejandro Colomar <alx@kernel.org>