wctrans.3, wctype.3: srcfix
Convert from unfilled text using `nf` and `fi` requests to `IP` macro
calls and tbl(1) tables.

This change increases the item indentation slightly, as I elected not to
specify one in the `IP` calls.  The content still fits easily in an
80-column terminal.

The reason for this change is to make the man page cross references
susceptible to scripted rewriting (and ultimately to make them
hyperlinkable).  See, e.g.,
<https://lore.kernel.org/linux-man/20240831182027.b6pduwkthk5b3tcf@illithid/>.

Signed-off-by: "G. Branden Robinson" <g.branden.robinson@gmail.com>
Message-ID: <20240901032513.afty634vpnhe24zi@illithid>
Signed-off-by: Alejandro Colomar <alx@kernel.org>