bind.2: ERRORS: Document possible errors from protocol
When looking through the errors of socket(2) I noticed that it specifies
the selected underlying protocol may extend the potential errors
returned.  For example, using AF_PACKET and SOCK_RAW can return EPERM if
the user does not have CAP_NET_RAW or uid 0 (this is all fully
documented).

However, AF_PACKET and SOCK_RAW extend the potential errors returned
from bind(2) as well.  For example, calling bind(2) with an invalid
sll_ifindex set on the sock_addr passed in will return ENODEV.

While this possibility is documented in the raw(7) manual page, the
bind(2) manual page does not mention that its potential set of errors
can be extended by the underlying protocol.  This patch simply
duplicates the relevant language from the socket(2) manual page to the
bind(2) manual page.

It is possible further extensions for send(2), recv(2), setsockopt(2),
etc. are also undocumented, but I have not yet verified this.

Signed-off-by: Alejandro Colomar <alx@kernel.org>