scripts/bash_aliases: man_gitstaged(): Trim the dirname(1) only for files within man/
I changed the behavior at some point to trim the dirname(1) for every
file.  However, that was mainly due to the inconvenience of not having a
man/ directory.  Also, I haven't been consistent with that behavior, and
have been manually adding back the dirname(1), so let's bring back the
old behavior --which is BTW still what the comment says it does--.

Signed-off-by: Alejandro Colomar <alx@kernel.org>