printf.3: wfix
Improve terminology in %a description

The term "decimal point" does not technically apply when using bases
other than 10; the more generic term is "radix point".  Update the
description of the a/A conversion specifier (i.e., for hexadecimal
floating point output) in printf(3) to use this terminology.

I do note that POSIX.1-2024 [1] does use the term "decimal-point
character" here, but I still maintain that using "radix point" is a
better term for that object in the %a description.  (Confusingly, POSIX
does refer to "radix character" in the descriptions of %f and %e, where
reference to "decimal" instead of "radix" would actually make sense.)

Link: [1] <https://pubs.opengroup.org/onlinepubs/9799919799/functions/printf.html>
Signed-off-by: Ken Pizzini <ken@gnu.org>
Message-ID: <db91cc6f-93cc-4e99-806c-7a8b86232848@vagg4fs7.msa.explicate.org>
Signed-off-by: Alejandro Colomar <alx@kernel.org>