man/: Terminology consistency reforms (n, size, length)
Use 'length' for the lenght of a string.
Use 'n' for the number of elements.
Use 'size' for the number of bytes.  (And in wide-character string
functions, 'size' also refers to the number of wide characters.)

The change is quite large, and I might have made some mistakes.
But overall, this should improve consistency in use of these terms.

Signed-off-by: Alejandro Colomar <alx@kernel.org>