man/man2/fork.2: Add _exit(2) to example
The _exit(2) function is a better choice for exiting a child in many
cases.  Most prominently it avoids calls of functions registered with
atexit(3) by the parent.

There are valid reasons to call exit(3) and the example is actually one
of them: flush FILE-based output.  Since atexit(3) is never called, we
could just stay with exit(3).

Signed-off-by: Tobias Stoeckmann <tobias@stoeckmann.org>
Message-ID: <tngwcffbrzbfkj6vrxgxpekrp3bzuftdy2mzow56xyfkrcna2w@nbgr2ourerxo>
Link: <https://github.com/shadow-maint/shadow/pull/1171>
Signed-off-by: Alejandro Colomar <alx@kernel.org>