man/man7/man-pages.7: Stop telling contributors to write titles in all caps
Recently, I submitted my first patch to the Linux man-pages project.  In
my patch, I had created a new manual page.  On the manual page’s title
line, I had written the title of my new page in all caps because
man-pages(7) said that I should write it that way.  It turns out that
man-pages(7) was wrong and that the title on the title line should have
matched the title in the manual page’s filename [1][2].  This commit
corrects man-pages(7) so that it does not tell contributors to use all
caps when writing titles on title lines.

Link: [1] <https://lore.kernel.org/linux-man/rph24kz36vysoeix4qoxxxcwq3c3fskws2vmxkkgcb2lpits3f@ysm7ug66svzh/T/#mc84250a6634d7f776118879021331001cceccbe5>
Link: [2] <https://lore.kernel.org/linux-man/mog6nnwzwl2dmlrec5b7l76wbxlsnklvdulok644x6o557trib@3zwtoz47r4x3/T/#mf71422d0e159210a7ca2798f2bba50a668e1410e>
Message-ID: <20250114211427.160509-1-jason@jasonyundt.email>
Signed-off-by: Alejandro Colomar <alx@kernel.org>